nlp-service/requirements.txt
flask==3.0.0
openai==1.3.5
python-dotenv==1.0.0